+++ Reqnroll/Bindings/Discovery/BindingSourceProcessor.cs
@@ -199,6 +199,7 @@ namespace Reqnroll.Bindings.Discovery
         {
             string regex = stepArgumentTransformationAttribute.TryGetAttributeValue<string>(0) ?? stepArgumentTransformationAttribute.TryGetAttributeValue<string>(nameof(StepArgumentTransformationAttribute.Regex));
             string name = stepArgumentTransformationAttribute.TryGetAttributeValue<string>(nameof(StepArgumentTransformationAttribute.Name));
+            int order = stepArgumentTransformationAttribute.TryGetAttributeValue<int>(nameof(StepArgumentTransformationAttribute.Order));
 
             var validationResult = ValidateStepArgumentTransformation(bindingSourceMethod, stepArgumentTransformationAttribute);
             if (!validationResult.IsValid)
@@ -207,7 +208,7 @@ namespace Reqnroll.Bindings.Discovery
                 return;
             }
 
-            var stepArgumentTransformationBinding = _bindingFactory.CreateStepArgumentTransformation(regex, bindingSourceMethod.BindingMethod, name);
+            var stepArgumentTransformationBinding = _bindingFactory.CreateStepArgumentTransformation(regex, bindingSourceMethod.BindingMethod, name, order);
 
             ProcessStepArgumentTransformationBinding(stepArgumentTransformationBinding);
         }