+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Reqnroll.Verify: Support for Verify v24 (Verify.Xunit v24.2.0) for .NET 4.7.2+ and .NET 6.0+. For earlier versions of Verify or for .NET 4.6.2, use the latest 2.0.3 version of the plugin that is compatible with Reqnroll v2.*. (#151)
 * Optimize creation of test-thread context using test framework independent resource pooling (#144)
 * Support DateTimeOffset in value comparer (#180)
+* Support 'Order' parameter for `StepArgumentTransformationAttribute` to prioritize execution
 
 ## Bug fixes: